% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageTimestamp.R
\name{packageTimestamp}
\alias{packageTimestamp}
\title{packageTimestamp}
\usage{
packageTimestamp(pkgname)
}
\arguments{
\item{pkgname}{(character) package name}
}
\description{
Extract the "Built" timestamp from a package's description.
}
