% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Carefully View}
\usage{
view(..., careful = TRUE, max_rows = 5000, verbose = getOption("verbose"))
}
\arguments{
\item{max_rows}{(integer) prompt if viewing data with more than this many rows}

\item{verbose}{(logical)}

\item{x}{an \R object to view (maybe tabular data)}

\item{title}{title for viewer window}
}
\description{
Carefully View
}
\seealso{
View
}
