% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome_to.R
\name{welcome_to}
\alias{welcome_to}
\title{welcome_to}
\usage{
welcome_to(pkgname)
}
\arguments{
\item{pkgname}{(character) package name}
}
\description{
Print welcome message and package version/timestamp.
}
